Fixed most of build warnings and few bugs related to scripts and looting

+++ src/game/ChatCommands/PlayerStatsMods.cpp
@@ -1006,7 +1006,7 @@ bool ChatHandler::HandleModifyMoneyCommand(char* args)
         }
         else
         {
-            if (newmoney > MAX_MONEY_AMOUNT)
+            if (newmoney > int64(MAX_MONEY_AMOUNT))
             {
                 newmoney = MAX_MONEY_AMOUNT;
             }
@@ -1027,7 +1027,7 @@ bool ChatHandler::HandleModifyMoneyCommand(char* args)
             ChatHandler(chr).PSendSysMessage(LANG_YOURS_MONEY_GIVEN, GetNameLink().c_str(), MoneyToString(addmoney).c_str());
         }
 
-        if (addmoney >= MAX_MONEY_AMOUNT)
+        if (addmoney >= int64(MAX_MONEY_AMOUNT))
         {
             chr->SetMoney(MAX_MONEY_AMOUNT);
         }
