+++ src/game/WorldHandlers/SkillHandler.cpp
@@ -94,7 +94,7 @@ void WorldSession::HandleLearnPreviewTalents(WorldPacket& recvPacket)
     {
         if (TalentTabEntry const* talentTabEntry = sTalentTabStore.LookupEntry(_player->GetPrimaryTalentTree(_player->GetActiveSpec())))
         {
-            if (talentTabEntry->OrderIndex != tabPage)
+            if (int32(talentTabEntry->OrderIndex) != tabPage)
             {
                 recvPacket.rfinish();
                 sLog.outError("WorldSession::HandleLearnPreviewTalents: tabPage != talent tabPage for %s (account %u)", GetPlayerName(), GetAccountId());
