+++ src/game/WorldHandlers/SpellEffectHealPower.cpp
@@ -134,7 +134,7 @@ void Spell::EffectUnlearnSpecialization(SpellEffectEntry const* effect)
  */
 void Spell::EffectPowerDrain(SpellEffectEntry const* effect)
 {
-    if (effect->EffectMiscValue_0 < 0 || effect->EffectMiscValue_0 >= MAX_POWERS)
+    if (effect->EffectMiscValue_0 < 0 || effect->EffectMiscValue_0 >= int32(MAX_POWERS))
     {
         return;
     }
@@ -222,7 +222,7 @@ void Spell::EffectSendEvent(SpellEffectEntry const* effect)
  */
 void Spell::EffectPowerBurn(SpellEffectEntry const* effect)
 {
-    if (effect->EffectMiscValue_0 < 0 || effect->EffectMiscValue_0 >= MAX_POWERS)
+    if (effect->EffectMiscValue_0 < 0 || effect->EffectMiscValue_0 >= int32(MAX_POWERS))
     {
         return;
     }
@@ -753,7 +753,7 @@ void Spell::EffectEnergize(SpellEffectEntry const* effect)
         return;
     }
 
-    if (effect->EffectMiscValue_0 < 0 || effect->EffectMiscValue_0 >= MAX_POWERS)
+    if (effect->EffectMiscValue_0 < 0 || effect->EffectMiscValue_0 >= int32(MAX_POWERS))
     {
         return;
     }
@@ -877,7 +877,7 @@ void Spell::EffectEnergisePct(SpellEffectEntry const* effect)
         return;
     }
 
-    if (effect->EffectMiscValue_0 < 0 || effect->EffectMiscValue_0 >= MAX_POWERS)
+    if (effect->EffectMiscValue_0 < 0 || effect->EffectMiscValue_0 >= int32(MAX_POWERS))
     {
         return;
     }
